Ensure components invoke `render` method.
